rename: game 페이지 생성 및 game 폴더 생성
