+++ FE/src/components/game/Bottombar.tsx
