+++ FE/src/components/game/ChattingItem.tsx
